+++ CHANGELOG.md
@@ -6,6 +6,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Added
+- Support for IP address as domain in RCPT TO (#285)
+
 
 ## [v3.0.3] - 2022-08-07